requirements.txt
anyio==4.11.0
asgiref==3.10.0
certifi==2025.11.12
Django==4.2.10
djangorestframework==3.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot==22.3.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
psycopg[binary]==3.1.19
python-dotenv==1.1.1
